+++ tybotRoute/models/TiledeskChatbotConst.js
@@ -1,19 +1,36 @@
 class TiledeskChatbotConst {
     
-    static REQ_DEPARTMENT_ID_KEY = "tdDepartmentId";
-    static REQ_PROJECT_ID_KEY = "tdProjectId";
-    static REQ_REQUEST_ID_KEY = "tdRequestId";
-    static REQ_CHATBOT_NAME_KEY = "tdChatbotName";
-    static REQ_LAST_USER_TEXT_KEY = "tdLastUserText";
-    static REQ_LAST_MESSAGE_ID_KEY = "tdLastMessageId";
-    static REQ_COUNTRY_KEY = "tdCountry";
-    static REQ_CITY_KEY = "tdCity";
-    static REQ_USER_SOURCE_PAGE_KEY = "tdUserSourcePage";
-    static REQ_USER_LANGUAGE_KEY = "tdUserLanguage";
-    static REQ_USER_AGENT_KEY = "tdUserAgent";
-    static REQ_DEPARTMENT_NAME_KEY = "tdDepartmentName";
-    static REQ_END_USER_ID_KEY = "tdEndUserId";
-    static REQ_END_USER_IP_ADDRESS_KEY = "tdEndUserIpAddress";
+    static REQ_DEPARTMENT_ID_KEY = "department_id";
+    static REQ_PROJECT_ID_KEY = "project_id";
+    static REQ_REQUEST_ID_KEY = "conversation_id";
+    static REQ_CHATBOT_NAME_KEY = "chatbot_name";
+    static REQ_LAST_USER_TEXT_KEY = "last_user_text";
+    static REQ_LAST_MESSAGE_ID_KEY = "last_message_id";
+    static REQ_COUNTRY_KEY = "user_country";
+    static REQ_CITY_KEY = "user_city";
+    static REQ_USER_SOURCE_PAGE_KEY = "user_source_page";
+    static REQ_USER_LANGUAGE_KEY = "user_language";
+    static REQ_USER_AGENT_KEY = "user_agent";
+    static REQ_DEPARTMENT_NAME_KEY = "department_name";
+    static REQ_END_USER_ID_KEY = "user_id";
+    static REQ_END_USER_IP_ADDRESS_KEY = "user_ip_address";
+    static REQ_CHAT_URL = "chat_url";
+
+    // static REQ_DEPARTMENT_ID_KEY = "tdDepartmentId";
+    // static REQ_PROJECT_ID_KEY = "projectId";
+    // static REQ_REQUEST_ID_KEY = "tdRequestId";
+    // static REQ_CHATBOT_NAME_KEY = "tdChatbotName";
+    // static REQ_LAST_USER_TEXT_KEY = "tdLastUserText";
+    // static REQ_LAST_MESSAGE_ID_KEY = "tdLastMessageId";
+    // static REQ_COUNTRY_KEY = "tdCountry";
+    // static REQ_CITY_KEY = "tdCity";
+    // static REQ_USER_SOURCE_PAGE_KEY = "tdUserSourcePage";
+    // static REQ_USER_LANGUAGE_KEY = "tdUserLanguage";
+    // static REQ_USER_AGENT_KEY = "tdUserAgent";
+    // static REQ_DEPARTMENT_NAME_KEY = "tdDepartmentName";
+    // static REQ_END_USER_ID_KEY = "tdEndUserId";
+    // static REQ_END_USER_IP_ADDRESS_KEY = "tdEndUserIpAddress";
+    // static REQ_CHAT_URL = "chat_url";
     
 }