+++ tybotRoute/test/condition_json_to_expression_test.js
@@ -439,7 +439,107 @@ describe('JSON to expression', function() {
     assert(result === true);
   });
 
- 
+  it('validates variable names', async () => {
+    let is_valid;
+    is_valid = TiledeskExpression.validateVariableName("1");
+    assert(is_valid === false);
+  });
+
+  // TEST CONDITIONS WITH INVALID OPERANDS
+
+  it('test condition with invalid variable operand', async () => {
+    const condition = {
+      "type": "condition",
+      "operand1": "age()", // invalid chars ()
+      "operator": TiledeskExpression.OPERATORS.lessThan.name,
+      "operand2": {
+        type: "const",
+        value: "10"
+      }
+    }
+    const expression = TiledeskExpression.JSONConditionToExpression(condition);
+    console.log("expression:", expression);
+    assert(expression === null);
+  });
+
+  it('test condition with invalid variable operand', async () => {
+    const condition = {
+      "type": "condition",
+      "operand1": "age",
+      "operator": TiledeskExpression.OPERATORS.lessThan.name,
+      "operand2": {
+        type: "var",
+        name: "12" // invalid chars: starts with numbers
+      }
+    }
+    const expression = TiledeskExpression.JSONConditionToExpression(condition);
+    console.log("invalid expression:", expression);
+    assert(expression === null);
+  });
+
+  it('test multiple groups with invalid variable names', async () => {
+    const part1 = {
+      "type": "condition",
+      "operand1": "11height", // invalid, starts with numbers
+      "operator": TiledeskExpression.OPERATORS.greaterThan.name,
+      "operand2": {
+        type: "const",
+        value: "1"
+      }
+    }
+    const part2 = {
+      "type": "operator",
+      "operatorName": TiledeskExpression.OPERATORS.AND.name
+    }
+    const part3 = {
+      "type": "condition",
+      "operand1": "name",
+      "operator": TiledeskExpression.OPERATORS.startsWith.name,
+      "operand2": {
+        type: "const",
+        value: "And"
+      }
+    }
+    const part5 = {
+      "type": "condition",
+      "operand1": "size",
+      "operator": TiledeskExpression.OPERATORS.equalAsNumbers.name,
+      "operand2": {
+        type: "const",
+        value: "03"
+      }
+    }
+    const part6 = {
+      "type": "operator",
+      "operatorName": TiledeskExpression.OPERATORS.AND.name
+    }
+    const part7 = {
+      "type": "condition",
+      "operand1": "name",
+      "operator": TiledeskExpression.OPERATORS.equalAsStrings.name,
+      "operand2": {
+        type: "const",
+        value: "Andrea"
+      }
+    }
+    const group1 = {
+      type: "expression",
+      conditions: [ part1, part2, part3 ]
+    }
+    const group_operator = {
+      "type": "operator",
+      "operatorName": TiledeskExpression.OPERATORS.OR.name
+    }
+    const group2 = {
+      type: "expression",
+      conditions: [ part5, part6, part7 ]
+    }
+    const groups = [group1, group_operator, group2];
+    const expression = TiledeskExpression.JSONGroupsToExpression(groups);
+    console.log("invalid group expression:", expression);
+    assert(expression === null);
+  });
+
   // Struct
     // const expression = {
     //   groups: [